# sgrti-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SGRTI API",
        "description": "Core Transaccional",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/v1/core/requirements": {
            "get": {
                "tags": [
                    "Requirements"
                ],
                "summary": "Obtener lista de requerimientos",
                "operationId": "c37c89f4617e77cdac0ce7f070ae52da",
                "responses": {
                    "200": {
                        "description": "Lista recuperada exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Requirements"
                ],
                "summary": "Crear un nuevo requerimiento",
                "description": "Momento 1: Registro Inicial de Requerimiento (Responsabilidad: Coordinador)\nCU-004: Crear Requerimiento\nCU-012: Gestionar Unidad Solicitante\n* El parámetro StoreRequirementRequest valida automáticamente los datos\nantes de entrar a este método.\n* @param StoreRequirementRequest $request",
                "operationId": "365fe5b43581dd5ded73db3d6938fdc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "priority"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Falla en nodo central"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Se requiere revisión de fibra"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "ALTA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Requerimiento creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/core/requirements/{id}": {
            "get": {
                "tags": [
                    "Requirements"
                ],
                "summary": "Consultar detalle de un requerimiento",
                "description": "Mostrar el detalle de un requerimiento específico.",
                "operationId": "434958cec9644f5e2da051ca4fe70e4d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID o ID del requerimiento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del requerimiento encontrado"
                    },
                    "404": {
                        "description": "Requerimiento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Requirements"
                ],
                "summary": "Actualizar requerimiento (Sujeto a bloqueo ATF)",
                "operationId": "04231add821dec6908ce29be52345bed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del requerimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requerimiento actualizado"
                    },
                    "403": {
                        "description": "Acceso denegado: Requerimiento bloqueado por fase ATF"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Requirements"
                ],
                "summary": "Eliminar requerimiento con clave especial",
                "operationId": "6a30bb9158ddcf387449926bb4cacc41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "special_key",
                        "in": "query",
                        "description": "Clave de Operaciones Especiales",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado con éxito"
                    },
                    "401": {
                        "description": "Clave de seguridad incorrecta"
                    }
                }
            }
        },
        "/api/v1/core/requirements/{id}/estimation": {
            "patch": {
                "tags": [
                    "Requirements"
                ],
                "summary": "Actualizar datos de estimación técnica",
                "description": "Momento 2: Registro de Estimación (Responsabilidad: Consultor CSPE)\nCU-011: Estimar Requerimiento",
                "operationId": "0b8c703e14d3e3f9f1e9bac112fb12d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "horas_estimadas": {
                                        "type": "integer",
                                        "example": 40
                                    },
                                    "costo_estimado": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "tecnico_asignado": {
                                        "type": "string",
                                        "example": "Ing. Juan Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estimación actualizada correctamente"
                    },
                    "422": {
                        "description": "Datos de estimación inválidos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Requirements",
            "description": "Requirements"
        }
    ]
}